add plotting to pers landscape visual file
